fix: cleanup code for RTB
