+++ clickhouse/init/temperature.sql
@@ -1,3 +1,8 @@
+CREATE TABLE sensors.temperature_kafka
+(
+    data String
+) ENGINE = Kafka('redpanda:9092', 'temperature', 'ch_group_1', 'JSONAsString');
+
 CREATE TABLE sensors.temperatures
 (
     sensor_uuid UUID,
@@ -8,6 +13,15 @@ CREATE TABLE sensors.temperatures
     longitude   Float64
 ) ENGINE = MergeTree()
       ORDER BY (sensor_uuid, timestamp);
+
+CREATE MATERIALIZED VIEW sensors.temperature_topic_mv TO sensors.temperature as
+SELECT JSONExtractString(data, 'sensor_name')                AS sensor_name,
+       toUUID(JSONExtractString(data, 'sensor_uuid'))        AS sensor_uuid,
+       toDateTime64(JSONExtractString(data, 'timestamp'), 0) AS timestamp,
+       JSONExtractFloat(data, 'value')                       AS value,
+       JSONExtractFloat(data, 'latitude')                    AS latitude,
+       JSONExtractFloat(data, 'longitude')                   AS longitude
+FROM sensors.temperature_kafka;
 
 -- Real-time
 CREATE TABLE sensors.temperatures_realtime
