--- connectors/configs/clickhouse.json
@@ -1,34 +1,0 @@
-{
-    "name": "clickhouse-sink",
-    "config": {
-        "connector.class": "com.clickhouse.kafka.connect.ClickHouseSinkConnector",
-        "name": "clickhouse-sink",
-        "key.converter": "org.apache.kafka.connect.storage.StringConverter",
-        "topics": "temperature,traffic,ecological_island",
-        "value.converter": "io.confluent.connect.avro.AvroConverter",
-        "value.converter.schemas.enable": "true",
-        "value.converter.schema.registry.url": "http://redpanda:8081",
-        "value.subject.name.strategy": "io.confluent.kafka.serializers.subject.RecordNameStrategy",
-        "header.converter": "org.apache.kafka.connect.storage.SimpleHeaderConverter",
-        "hostname": "clickhouse",
-        "port": "8123",
-        "database": "sensors",
-        "username": "admin",
-        "password": "admin",
-        "ssl": false,
-        "exactlyOnce": false,
-        "suppressTableExistenceException": false,
-        "topic2TableMap": "temperature=temperatures,traffic=traffic,ecological_island=ecological_island",
-        "customInsertFormat": false,
-        "enableDbTopicSplit": false,
-        "errors.log.enable": false,
-        "errors.log.include.messages": false,
-        "errors.deadletterqueue.context.headers.enable": false,
-        "errors.tolerance": "none",
-        "transforms": "TimestampConverter",
-        "transforms.TimestampConverter.type": "org.apache.kafka.connect.transforms.TimestampConverter$Value",
-        "transforms.TimestampConverter.format": "yyyy-MM-dd'T'HH:mm:ss",
-        "transforms.TimestampConverter.field": "timestamp",
-        "transforms.TimestampConverter.target.type": "Timestamp"
-    }
-}
