--- connectors/plugins/avro_converter/doc/version.txt
@@ -1,2 +1,0 @@
-##This file must be updated at package time to have a valid package##
-invalid
