--- connectors/plugins/avro_converter/manifest.json
@@ -1,39 +1,0 @@
-{
-  "name" : "kafka-connect-avro-converter",
-  "version" : "7.6.1",
-  "title" : "Kafka Connect Avro Converter",
-  "description" : "The Kafka Connect Avro Converter integrates with <a href=\"https://docs.confluent.io/current/schema-registry/docs/intro.html\">Schema Registry</a> to convert data for Kafka Connect to and from Avro format.",
-  "owner" : {
-    "username" : "confluentinc",
-    "type" : "organization",
-    "name" : "Confluent, Inc.",
-    "url" : "https://confluent.io/",
-    "logo" : "assets/confluent.png"
-  },
-  "support" : {
-    "summary" : "Confluent supports the Avro Converter alongside community members as part of its Confluent Platform offering.",
-    "url" : "https://docs.confluent.io/current/",
-    "logo" : "assets/confluent.png",
-    "provider_name" : "Confluent, Inc."
-  },
-  "tags" : [ "schema registry", "avro" ],
-  "features" : {
-    "supported_encodings" : [ "avro" ],
-    "single_message_transforms" : false,
-    "confluent_control_center_integration" : false,
-    "kafka_connect_api" : false
-  },
-  "documentation_url" : "https://docs.confluent.io/current/schema-registry/docs/connect.html",
-  "source_url" : "https://github.com/confluentinc/schema-registry",
-  "docker_image" : {
-    "namespace" : "confluentinc",
-    "name" : "cp-kafka-connect",
-    "tag" : "7.6.1"
-  },
-  "license" : [ {
-    "name" : "Apache License 2.0",
-    "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
-  } ],
-  "component_types" : [ "converter" ],
-  "release_date" : "2024-04-10"
-}
