--- connectors/plugins/clickhouse/doc/README.md
@@ -1,15 +1,0 @@
-# ClickHouse Kafka Connect Sink
-
-## About
-clickhouse-kafka-connect is the official Kafka Connect sink connector for [ClickHouse](https://clickhouse.com/).
-
-The Kafka connector delivers data from a Kafka topic to a ClickHouse table.
-## Documentation
-
-See the [ClickHouse website](https://clickhouse.com/docs/en/integrations/kafka/clickhouse-kafka-connect-sink) for the full documentation entry.
-
-## Design
-For a full overview of the design and how exactly-once delivery semantics are achieved, see the [design document](./docs/DESIGN.md).
-
-## Help
-For additional help, please [file an issue in the repository](https://github.com/ClickHouse/clickhouse-kafka-connect/issues) or raise a question in [ClickHouse public Slack](https://clickhouse.com/slack).
