+++ simulator/src/producers/kafka_producer.py
@@ -1,27 +1,28 @@
 import logging as log
+import json
 
 import kafka
 
-from .producer_strategy import ProducerStrategy
 from ..models.raw_data.raw_data import RawData
-from ..serializers.serializer_strategy import SerializerStrategy
+from ..serializers.json_serializer import JSONSerializer
 
 
-class KafkaProducer(ProducerStrategy):
+class KafkaProducer():
 
-    def __init__(self, *, serializer: SerializerStrategy, bootstrap_servers: list[str],
+    def __init__(self, *, serializer: JSONSerializer, bootstrap_servers: list[str],
                  max_block_ms: int, acks: int) -> None:
-        super().__init__(serializer)
+        self.serializer = serializer
         self._producer = kafka.KafkaProducer(
             bootstrap_servers=bootstrap_servers,
             max_block_ms=max_block_ms,
+            value_serializer=lambda v: json.dumps(v).encode("utf-8"),
             acks=acks,
         )
 
     def produce(self, data: RawData) -> None:
-        serialized = self._serializer.serialize(data)
+        serialized = data.accept(self.serializer)
         log.info(f'Producing data to topic {data.topic}: {data}')
-        self._producer.send(data.topic, value=serialized)
+        self._producer.send(data.topic, serialized)
         self._producer.flush()
 
     def close(self) -> None:
