--- simulator/src/schemas/temperature-value.avsc
@@ -1,30 +1,0 @@
-{
-  "type": "record",
-  "name": "Temperature",
-  "fields": [
-    {
-      "name": "sensor_uuid",
-      "type": "string"
-    },
-    {
-      "name": "sensor_name",
-      "type": "string"
-    },
-    {
-      "name": "latitude",
-      "type": "double"
-    },
-    {
-      "name": "longitude",
-      "type": "double"
-    },
-    {
-      "name": "timestamp",
-      "type": "string"
-    },
-    {
-      "name": "value",
-      "type": "float"
-    }
-  ]
-}
